pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "minicnn"
version = "0.7.0"
description = "GitHub-ready configuration-driven MiniCNN with flexible registries and native CUDA migration path"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
torch = ["torch>=2.0"]
dev = ["pytest>=8.0"]

[project.scripts]
minicnn = "minicnn.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
